Add android and ios example
