+++ example/package.json
@@ -8,7 +8,7 @@
   },
   "dependencies": {
     "react": "16.0.0-alpha.6",
-    "react-native": "0.44.0",
+    "react-native": "0.43.0",
     "react-native-masonry": "^0.1.0"
   },
   "devDependencies": {
